Revert "new schematron node service."
This reverts commit c540920938969f60de5f8ed5bcdcd9ceba099e14.
