+++ ccdaservice/package.json
@@ -12,11 +12,6 @@
         "lodash": "4.17.21",
         "net": "1.0.2",
         "uuid": "8.3.2",
-        "xmljson": "0.2.0",
-        "body-parser": "^1.16.0",
-        "cda-schematron": "^1.0.0",
-        "express": "^4.14.0",
-        "morgan": "^1.7.0",
-        "winston": "^2.3.1"
+        "xmljson": "0.2.0"
     }
 }
